add product crud

+++ database/migrations/2025_08_06_161903_create_categories_table.php
@@ -24,6 +24,8 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('categories');
+        if (! app()->isProduction()) {
+            Schema::dropIfExists('categories');
+        }
     }
 };
