+++ database/migrations/2025_08_06_161927_create_params_table.php
@@ -24,6 +24,8 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('params');
+        if (! app()->isProduction()) {
+            Schema::dropIfExists('params');
+        }
     }
 };
