+++ database/migrations/2025_08_06_162202_create_param_product_table.php
@@ -25,6 +25,8 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('param_product');
+        if (! app()->isProduction()) {
+            Schema::dropIfExists('param_product');
+        }
     }
 };
