+++ database/migrations/2025_08_06_162241_create_orders_table.php
@@ -25,6 +25,8 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('orders');
+        if (! app()->isProduction()) {
+            Schema::dropIfExists('orders');
+        }
     }
 };
