+++ database/migrations/2025_08_06_162244_create_product_user_cards_table.php
@@ -27,6 +27,8 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('product_user_cards');
+        if (! app()->isProduction()) {
+            Schema::dropIfExists('product_user_cards');
+        }
     }
 };
