+++ database/migrations/2025_08_06_162341_create_images_table.php
@@ -24,6 +24,8 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('images');
+        if (! app()->isProduction()) {
+            Schema::dropIfExists('images');
+        }
     }
 };
