+++ database/migrations/2025_08_10_154707_create_roles_table.php
@@ -23,6 +23,8 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('roles');
+        if (! app()->isProduction()) {
+            Schema::dropIfExists('roles');
+        }
     }
 };
