+++ database/migrations/2025_08_10_154719_create_role_user_table.php
@@ -24,6 +24,8 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('role_user');
+        if (! app()->isProduction()) {
+            Schema::dropIfExists('role_user');
+        }
     }
 };
